Converted the storage to use longs instead of doubles

+++ src/main/java/com/ofek2608/deep_pocket/api/DeepPocketClientApi.java
@@ -27,5 +27,5 @@ public interface DeepPocketClientApi extends DeepPocketApi {
 	void setPermitPublicPocket(boolean value);
 
 	PlayerKnowledge getKnowledge();
-	Stream<Map.Entry<ItemType,Double>> getSortedKnowledge(Pocket pocket);
+	Stream<Map.Entry<ItemType,Long>> getSortedKnowledge(Pocket pocket);
 }
